Reverted previous `view` changes; implemented ZappaFS. (#7)
There's little chance we'll ever get async templates with parameterized `read` function.
So we're monkey-patching the `fs` module so that we always get the right answer.

Note: zappajs must be `require`d before any module that might need to access inline views (such as Express, Connect, or a template engine that supports partials).