Removed old examples.
Mostly I don't want to have to track Express to make sure the comparison holds.
The .js example would have to be rewritten to use `this` everywhere, but contrarily to early zappa versions (which modified the scope) current zappa code doesn't do that anymore so the comparison with plain JS is easier.