Removed `@stylus`; added `@with` to allow in-code Zappa changes.
From an API perspective, it makes no sense to mix the code that creates the server (host, port, https) with the semantics of the callback.
The callback should be self-contained, and if it requires `stylus`, `less`, `eco` etc. it should be able to request them on its own, without creating a dependency on how it is called.
Hence the new `@with` API, which replaces the old `(require_)css` option of `zappa.run()`.

Regarding the removal of `@stylus` the rationale is that coffee-css is the default styling engine due to plain coffee-script inclusion. (There's no ambiguity when you're calling `@css` whether we're dealing with plain CSS or coffee-css.)
`stylus` no longer has a preferred status, it should be declared like any other stylesheet template engine.