Actually call the building script. (Fixes #112.)
This is overly convoluted and I get confused every time, but essentially the call to `p_fun` still returns the original function, and that function must be called for `window.zappa` etc. to be created.
And since `p_fun` creates a statement instead of an expression, we must use it to store in a variable.